임시로 만들어논 닥터 컨트롤러에서 약사 컨트롤러로 ㅇ옮겼음
