+++ src/main/java/mna/homespital/controller/DoctorController.java
@@ -23,42 +23,6 @@ import org.springframework.web.servlet.ModelAndView;
 @RequestMapping("/")
 public class DoctorController {
 
-    @Autowired
-    HttpSession session;
-
-    @Autowired
-    PharService pharService;
-
-    //나의진료내역 (인성)
-    @GetMapping("/customerList")
-    public String customerList(HttpSession session, Model m) throws Exception {
-        try {
-//            String email = (String) session.getAttribute("email");
-//            String name = (String) session.getAttribute("name");
-//            int searchNumber = pharService.getNumberByEmail(email);
-            Diagnosis diagnosis = new Diagnosis();
-//            diagnosis.setUser_number(searchNumber);
-            m.addAttribute("diagnosis", diagnosis);
-        } catch (Exception e) {
-            System.out.println("Catch() join");
-            e.printStackTrace();
-            return "common/err";
-        }
-        return "admin/phar/customerList";
-    }
-
-    //진료내역 리스트 출력 (인성)
-    @ResponseBody
-    @GetMapping("/pharCustomerRecordsList")
-    public ArrayList<HashMap<String, Object>> pharCustomerRecordsList(@RequestParam int phar_number) {
-        ArrayList<HashMap<String, Object>> customerList = new ArrayList<>();
-        try {
-            customerList = pharService.pharCustomerRecordsList(phar_number);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return customerList;
-    }
 
 }
