+++ src/main/java/mna/homespital/controller/PharmacyController.java
@@ -1,16 +1,21 @@
 package mna.homespital.controller;
 
+import mna.homespital.dto.Diagnosis;
 import mna.homespital.dto.Pharmacy;
 import mna.homespital.service.PharService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
+import java.util.ArrayList;
+import java.util.HashMap;
 
 @Controller
 public class PharmacyController {
@@ -19,7 +24,7 @@ public class PharmacyController {
     HttpSession session;
 
     @Autowired
-    PharService pharSerivce;
+    PharService pharService;
 
     @GetMapping("/pharmacyMainForm")
     public ModelAndView pharmacyMain() {
@@ -32,17 +37,17 @@ public class PharmacyController {
     }
 
     //용식:약사 로그인
-//    @PostMapping("PharmacyLogin.do")
-//    public String pharmacyLogin(@RequestParam("email") String email, @RequestParam("password") String password) {
-//        try {
-//            pharSerivce.login(email, password);
-//            session.setAttribute("email", email);
-//            return "redirect:/pharmacyMainForm";
-//        } catch (Exception e) {
-//            e.printStackTrace();
-//            return "redirect:/pharmacyLoginForm";
-//        }
-//    }
+    @PostMapping("PharmacyLogin.do")
+    public String pharmacyLogin(@RequestParam("email") String email, @RequestParam("password") String password) {
+        try {
+            pharService.pharmacyLogin(email, password);
+            session.setAttribute("email", email);
+            return "redirect:/pharmacyMainForm";
+        } catch (Exception e) {
+            e.printStackTrace();
+            return "redirect:/pharmacyLoginForm";
+        }
+    }
 
     //용식:약사 회원가입
     @PostMapping("pharmacyJoin.do")
@@ -61,7 +66,7 @@ public class PharmacyController {
         ModelAndView mv = new ModelAndView();
         try {
             System.out.println("여기들어오라ㅏ ㅇㅇㅇㅇㅇ");
-            pharSerivce.join(pharmacy);
+            pharService.join(pharmacy);
             mv.setViewName("redirect:/loginForm");
         } catch (Exception e) {
             e.printStackTrace();
@@ -69,5 +74,36 @@ public class PharmacyController {
         }
 
         return mv;
+    }
+
+    //나의진료내역 (인성)
+    @GetMapping("/customerList")
+    public String customerList(HttpSession session, Model m) throws Exception {
+        try {
+            String email = (String) session.getAttribute("email");
+            String name = (String) session.getAttribute("name");
+            int searchNumber = pharService.getNumberByEmail(email);
+            Diagnosis diagnosis = new Diagnosis();
+//            diagnosis.setUser_number(searchNumber);
+            m.addAttribute("diagnosis", diagnosis);
+        } catch (Exception e) {
+            System.out.println("Catch() join");
+            e.printStackTrace();
+            return "common/err";
+        }
+        return "admin/phar/customerList";
+    }
+
+    //진료내역 리스트 출력 (인성)
+    @ResponseBody
+    @GetMapping("/pharCustomerRecordsList")
+    public ArrayList<HashMap<String, Object>> pharCustomerRecordsList(@RequestParam int phar_number) {
+        ArrayList<HashMap<String, Object>> customerList = new ArrayList<>();
+        try {
+            customerList = pharService.pharCustomerRecordsList(phar_number);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return customerList;
     }
 }
